Move project export and export template manager into export folder
