+++ editor/editor_node.cpp
@@ -104,7 +104,8 @@
 #include "editor/editor_toaster.h"
 #include "editor/editor_translation_parser.h"
 #include "editor/export/editor_export.h"
-#include "editor/export_template_manager.h"
+#include "editor/export/export_template_manager.h"
+#include "editor/export/project_export.h"
 #include "editor/filesystem_dock.h"
 #include "editor/import/audio_stream_import_settings.h"
 #include "editor/import/dynamic_font_import_settings.h"
@@ -199,7 +200,6 @@
 #include "editor/plugins/visual_shader_editor_plugin.h"
 #include "editor/plugins/voxel_gi_editor_plugin.h"
 #include "editor/progress_dialog.h"
-#include "editor/project_export.h"
 #include "editor/project_settings_editor.h"
 #include "editor/register_exporters.h"
 #include "editor/scene_tree_dock.h"
